putting the broker back together
